+++ app/globals.css
@@ -1,5 +1,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Hind+Siliguri:wght@300;400;500;600;700&display=swap");
+
 @import "tailwindcss";
+
 @theme {
   --color-cream: #faf5f0;
   --font-sans: [ "Inter", "system-ui", "sans-serif"];
